more auto

+++ install.sh
@@ -7,6 +7,10 @@ elif [ "`uname`"x = "Darwin"x ];then
 	sed -i "" "s%PATH_VIM%${PATH_VIMCONFIG}%g" vimrc
 fi
 
-cp $HOME/.vimrc $HOME/.vimrc.old
-cp vimrc $HOME/.vimrc
-ln -s $PATH_VIMCONFIG ~/.vim
+if [ -f $HOME/.vimrc ];then
+	cp $HOME/.vimrc $HOME/.vimrc.old
+fi
+cp vimrc .vimrc
+cp .vimrc $HOME/.vimrc
+ln -s ${PATH_VIMCONFIG} ~/.vim
+
